+++ app/build.gradle.kts
@@ -114,9 +114,4 @@ dependencies {
     //Moshi
     implementation("com.squareup.moshi:moshi-kotlin:1.14.0")
 
-    //Retrofit
-    implementation("com.squareup.retrofit2:retrofit:2.9.0")
-    implementation("com.squareup.retrofit2:converter-gson:2.9.0")
-    implementation("com.squareup.retrofit2:converter-moshi:2.4.0")
-
 }
