--- app/src/main/java/com/example/novagincanabiblica/client/WordleService.kt
@@ -1,13 +1,0 @@
-package com.example.novagincanabiblica.client
-
-import com.example.novagincanabiblica.data.models.wordle.WordleCheck
-import retrofit2.Call
-import retrofit2.http.GET
-import retrofit2.http.Path
-
-interface WordleService {
-
-    @GET(value = "{word}")
-    fun checkWord(@Path(value = "word") word: String): Call<List<WordleCheck>>?
-
-}
