+++ app/src/main/java/com/example/novagincanabiblica/ui/screens/games/quiz/HowToPlayQuiz.kt
@@ -15,7 +15,7 @@ import com.example.novagincanabiblica.ui.theme.NovaGincanaBiblicaTheme
 fun HowToPlayQuizDialog() {
     BasicContainer {
         Column(modifier = Modifier.padding(16.dp), verticalArrangement = Arrangement.spacedBy(16.dp)) {
-            BasicText(text = "How to play The Bible Quiz?", fontSize = 22)
+            BasicText(text = "How to play The Bible Quiz?", fontSize = 28)
             BasicText(text = "In this game, you have 30 seconds to answer one question. You can choose 1 out of 4 answers. There are 4 difficulties in this game. They are:\n\n- Easy\n- Medium\n- Hard\n- Impossible\n\nIt is only one questions per day. Answer it wisely.")
         }
     }
