+++ app/src/main/java/com/example/novagincanabiblica/ui/screens/games/wordle/WordleStats.kt
@@ -19,21 +19,17 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.draw.shadow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.navigation.NavHostController
-import androidx.navigation.compose.rememberNavController
 import com.example.novagincanabiblica.data.models.WordleData
 import com.example.novagincanabiblica.data.models.WordleDataCalculated
 import com.example.novagincanabiblica.ui.basicviews.BasicContainer
 import com.example.novagincanabiblica.ui.basicviews.BasicText
-import com.example.novagincanabiblica.ui.basicviews.DaysStreak
+import com.example.novagincanabiblica.ui.screens.games.quiz.DaysStreak
 import com.example.novagincanabiblica.ui.basicviews.animateAlpha
 import com.example.novagincanabiblica.ui.basicviews.animateInt
-import com.example.novagincanabiblica.ui.screens.games.quiz.BackAndShare
+import com.example.novagincanabiblica.ui.screens.games.quiz.screens.BackAndShare
 import com.example.novagincanabiblica.ui.theme.NovaGincanaBiblicaTheme
-import com.example.novagincanabiblica.ui.theme.almostWhite
 import com.example.novagincanabiblica.ui.theme.closeToBlack
 import com.example.novagincanabiblica.ui.theme.correctAnswer
 import com.example.novagincanabiblica.ui.theme.lessWhite
