+++ app/src/main/java/com/example/novagincanabiblica/ui/theme/Color.kt
@@ -14,6 +14,7 @@ val lessWhite = Color(0xFFf0f0f0)
 val gray = Color(0xFFDDDDDD)
 val wrongPlace = Color(0xFFffdc73)
 val closeToBlack = Color(0xFF424242)
+val darkGray = Color(0xFF6B6B6B)
 val wrongAnswerSelected = Color(0xffef9a9a)
 val wrongAnswer = Color(0xffffcdd2)
 val wrongAnswerDark = Color(0xffc62828)
