+++ wordle.json
@@ -10,5 +10,17 @@
       "verse": "The Lord is my strength and my shield; My heart trusted in Him, and I am helped; Therefore my heart greatly rejoices, And with my song I will praise Him.",
       "word": "HEART"
     }
+  },
+  "pt": {
+    "day0": {
+      "reference": "Josué 1:9",
+      "verse": "Não fui eu que ordenei a você? Seja forte e corajoso! Não se apavore nem desanime, pois o Senhor, o seu Deus, estará com você por onde você andar.",
+      "word": "SENHOR"
+    },
+    "day1": {
+      "reference": "João 16:33",
+      "verse": "Eu disse essas coisas para que em mim vocês tenham paz. Neste mundo vocês terão aflições; contudo, tenham ânimo! Eu venci o mundo.",
+      "word": "ANIMO"
+    }
   }
 }
